image-classifier/requirements.txt
Flask==2.0.3
flask-cors==3.0.10
facenet-pytorch==2.5.2
torch==2.0.0
torchvision==0.15.1
numpy==1.23.0
Pillow==9.0.0
scikit-learn==1.6.0
joblib==1.2.0
gunicorn==20.1.0
numpy==1.23.0
pillow==9.0.0
tensorflow==2.13.0

